Create README.md
In summary, a tic tac toe game can be created using HTML, CSS, and JavaScript. The game board can be created using a container div with three rows of three cells each. CSS can be used to style the game board with a border, margins, and font size. JavaScript can be used to add event listeners to each cell that listens for clicks and updates the inner text of the cell to indicate which player has made a move. The game logic can be implemented using JavaScript, with code to check for winning combinations and end the game if necessary. Finally, the game can be reset and a new game started when the game is over.


HTML:
To start, you'll need to create the basic structure of the game board using HTML. You can create a div with a class of game-container to hold the game board. Inside the container, create three divs with a class of game-row to represent the rows of the game board. Inside each row div, create three divs with a class of game-cell to represent the individual cells of the game board. Here's what the HTML structure looks like:

CSS:
Next, you'll need to add some CSS to style the game board. You can use flexbox to create a responsive layout and set the width and height of the game cells. You can also add some styling to the game cells, such as a border and a cursor to indicate that they can be clicked. Here's an example of what the CSS might look like:

JavaScript:
Finally, you'll need to add some JavaScript to handle the game logic. You can add an event listener to each game cell that listens for a click event. When a cell is clicked, you can update the inner text of the cell to indicate which player has made a move (X or O). You'll also need to add some code to check if there is a winning combination of cells, or if the game has resulted in a tie. If the game has ended, you can reset the board and start a new game. Here's an example of what the JavaScript code might look like: